requirements.txt
crewai==0.28.0
langchain-openai==0.0.2.post1
crewai-tools==0.1.29
python-dotenv==1.0.0
argparse==1.4.0
pydantic==2.4.2
streamlit==1.30.0
plotly==5.18.0
pandas==2.1.3
pillow==10.1.0 